@@ -2,8 +2,7 @@ from __future__ import annotations
 
 import typing
 import warnings
-from packaging.version import Version
-from typing import Any, Callable, Dict, Optional
+from typing import Any
 
 import numpy as np
 import pandas as pd
@@ -34,16 +33,16 @@ from .base import is_geometry_type
 
 if typing.TYPE_CHECKING:
     import os
+    from collections.abc import Callable
 
 
 def _geoseries_constructor_with_fallback(
-    data=None, index=None, crs: Optional[Any] = None, **kwargs
+    data=None, index=None, crs: Any | None = None, **kwargs
 ):
-    """
-    A flexible constructor for GeoSeries._constructor, which needs to be able
+    """A flexible constructor for GeoSeries._constructor, which needs to be able
     to fall back to a Series (if a certain operation does not produce
-    geometries)
-    """
+    geometries).
+    """  # noqa: D401
     try:
         return GeoSeries(data=data, index=index, crs=crs, **kwargs)
     except TypeError:
@@ -97,7 +96,6 @@ class GeoSeries(GeoPandasBase, Series):
 
     Examples
     --------
-
     >>> from shapely.geometry import Point
     >>> s = geopandas.GeoSeries([Point(1, 1), Point(2, 2), Point(3, 3)])
     >>> s
@@ -154,7 +152,7 @@ class GeoSeries(GeoPandasBase, Series):
 
     """
 
-    def __init__(self, data=None, index=None, crs: Optional[Any] = None, **kwargs):
+    def __init__(self, data=None, index=None, crs: Any | None = None, **kwargs):
         if (
             hasattr(data, "crs")
             or (isinstance(data, pd.Series) and hasattr(data.array, "crs"))
@@ -164,7 +162,7 @@ class GeoSeries(GeoPandasBase, Series):
                 # make a copy to avoid setting CRS to passed GeometryArray
                 data = data.copy()
             else:
-                if not data.crs == crs:
+                if not data_crs == crs:
                     raise ValueError(
                         "CRS mismatch between CRS of the passed geometries "
                         "and 'crs'. Use 'GeoSeries.set_crs(crs, "
@@ -233,8 +231,17 @@ class GeoSeries(GeoPandasBase, Series):
         if not self.crs:
             self.crs = crs
 
-    def append(self, *args, **kwargs) -> GeoSeries:
-        return self._wrapped_pandas_method("append", *args, **kwargs)
+    @GeoPandasBase.crs.setter
+    def crs(self, value):
+        if self.crs is not None:
+            warnings.warn(
+                "Overriding the CRS of a GeoSeries that already has CRS. "
+                "This unsafe behavior will be deprecated in future versions. "
+                "Use GeoSeries.set_crs method instead.",
+                stacklevel=2,
+                category=FutureWarning,
+            )
+        self.geometry.values.crs = value
 
     @property
     def geometry(self) -> GeoSeries:
@@ -242,7 +249,7 @@ class GeoSeries(GeoPandasBase, Series):
 
     @property
     def x(self) -> Series:
-        """Return the x location of point geometries in a GeoSeries
+        """Return the x location of point geometries in a GeoSeries.
 
         Returns
         -------
@@ -250,7 +257,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> from shapely.geometry import Point
         >>> s = geopandas.GeoSeries([Point(1, 1), Point(2, 2), Point(3, 3)])
         >>> s.x
@@ -261,7 +267,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         See Also
         --------
-
         GeoSeries.y
         GeoSeries.z
 
@@ -270,7 +275,7 @@ class GeoSeries(GeoPandasBase, Series):
 
     @property
     def y(self) -> Series:
-        """Return the y location of point geometries in a GeoSeries
+        """Return the y location of point geometries in a GeoSeries.
 
         Returns
         -------
@@ -278,7 +283,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> from shapely.geometry import Point
         >>> s = geopandas.GeoSeries([Point(1, 1), Point(2, 2), Point(3, 3)])
         >>> s.y
@@ -289,16 +293,16 @@ class GeoSeries(GeoPandasBase, Series):
 
         See Also
         --------
-
         GeoSeries.x
         GeoSeries.z
+        GeoSeries.m
 
         """
         return _delegate_property("y", self)
 
     @property
     def z(self) -> Series:
-        """Return the z location of point geometries in a GeoSeries
+        """Return the z location of point geometries in a GeoSeries.
 
         Returns
         -------
@@ -306,7 +310,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> from shapely.geometry import Point
         >>> s = geopandas.GeoSeries([Point(1, 1, 1), Point(2, 2, 2), Point(3, 3, 3)])
         >>> s.z
@@ -317,12 +320,52 @@ class GeoSeries(GeoPandasBase, Series):
 
         See Also
         --------
-
         GeoSeries.x
         GeoSeries.y
+        GeoSeries.m
 
         """
         return _delegate_property("z", self)
+
+    @property
+    def m(self) -> Series:
+        """Return the m coordinate of point geometries in a GeoSeries.
+
+        Requires Shapely >= 2.1.
+
+        .. versionadded:: 1.1.0
+
+        Returns
+        -------
+        pandas.Series
+
+        Examples
+        --------
+        >>> from shapely.geometry import Point
+        >>> s = geopandas.GeoSeries.from_wkt(
+        ...     [
+        ...         "POINT M (2 3 5)",
+        ...         "POINT M (1 2 3)",
+        ...     ]
+        ... )
+        >>> s
+        0    POINT M (2 3 5)
+        1    POINT M (1 2 3)
+        dtype: geometry
+
+        >>> s.m
+        0    5.0
+        1    3.0
+        dtype: float64
+
+        See Also
+        --------
+        GeoSeries.x
+        GeoSeries.y
+        GeoSeries.z
+
+        """
+        return _delegate_property("m", self)
 
     @classmethod
     def from_file(cls, filename: os.PathLike | typing.IO, **kwargs) -> GeoSeries:
@@ -369,11 +412,10 @@ class GeoSeries(GeoPandasBase, Series):
 
     @classmethod
     def from_wkb(
-        cls, data, index=None, crs: Optional[Any] = None, on_invalid="raise", **kwargs
+        cls, data, index=None, crs: Any | None = None, on_invalid="raise", **kwargs
     ) -> GeoSeries:
-        """
-        Alternate constructor to create a ``GeoSeries``
-        from a list or array of WKB objects
+        r"""Alternate constructor to create a ``GeoSeries``
+        from a list or array of WKB objects.
 
         Parameters
         ----------
@@ -391,6 +433,9 @@ class GeoSeries(GeoPandasBase, Series):
             - warn: a warning will be raised and invalid WKB geometries will be returned
               as None.
             - ignore: invalid WKB geometries will be returned as None without a warning.
+            - fix: an effort is made to fix invalid input geometries (e.g. close
+              unclosed rings). If this is not possible, they are returned as ``None``
+              without a warning. Requires GEOS >= 3.11 and shapely >= 2.1.
 
         kwargs
             Additional arguments passed to the Series constructor,
@@ -404,6 +449,28 @@ class GeoSeries(GeoPandasBase, Series):
         --------
         GeoSeries.from_wkt
 
+        Examples
+        --------
+        >>> wkbs = [
+        ... (
+        ...     b"\x01\x01\x00\x00\x00\x00\x00\x00\x00"
+        ...     b"\x00\x00\xf0?\x00\x00\x00\x00\x00\x00\xf0?"
+        ... ),
+        ... (
+        ...     b"\x01\x01\x00\x00\x00\x00\x00\x00\x00"
+        ...     b"\x00\x00\x00@\x00\x00\x00\x00\x00\x00\x00@"
+        ... ),
+        ... (
+        ...    b"\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00"
+        ...    b"\x00\x08@\x00\x00\x00\x00\x00\x00\x08@"
+        ... ),
+        ... ]
+        >>> s = geopandas.GeoSeries.from_wkb(wkbs)
+        >>> s
+        0    POINT (1 1)
+        1    POINT (2 2)
+        2    POINT (3 3)
+        dtype: geometry
         """
         return cls._from_wkb_or_wkt(
             from_wkb, data, index=index, crs=crs, on_invalid=on_invalid, **kwargs
@@ -411,11 +478,10 @@ class GeoSeries(GeoPandasBase, Series):
 
     @classmethod
     def from_wkt(
-        cls, data, index=None, crs: Optional[Any] = None, on_invalid="raise", **kwargs
+        cls, data, index=None, crs: Any | None = None, on_invalid="raise", **kwargs
     ) -> GeoSeries:
-        """
-        Alternate constructor to create a ``GeoSeries``
-        from a list or array of WKT objects
+        """Alternate constructor to create a ``GeoSeries``
+        from a list or array of WKT objects.
 
         Parameters
         ----------
@@ -434,6 +500,9 @@ class GeoSeries(GeoPandasBase, Series):
               returned as ``None``.
             - ignore: invalid WKT geometries will be returned as ``None`` without a
               warning.
+            - fix: an effort is made to fix invalid input geometries (e.g. close
+              unclosed rings). If this is not possible, they are returned as ``None``
+              without a warning. Requires GEOS >= 3.11 and shapely >= 2.1.
 
         kwargs
             Additional arguments passed to the Series constructor,
@@ -449,7 +518,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> wkts = [
         ... 'POINT (1 1)',
         ... 'POINT (2 2)',
@@ -468,9 +536,8 @@ class GeoSeries(GeoPandasBase, Series):
 
     @classmethod
     def from_xy(cls, x, y, z=None, index=None, crs=None, **kwargs) -> GeoSeries:
-        """
-        Alternate constructor to create a :class:`~geopandas.GeoSeries` of Point
-        geometries from lists or arrays of x, y(, z) coordinates
+        """Alternate constructor to create a :class:`~geopandas.GeoSeries` of Point
+        geometries from lists or arrays of x, y(, z) coordinates.
 
         In case of geographic coordinates, it is assumed that longitude is captured
         by ``x`` coordinates and latitude by ``y``.
@@ -501,7 +568,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> x = [2.5, 5, -3.0]
         >>> y = [0.5, 1, 1.5]
         >>> s = geopandas.GeoSeries.from_xy(x, y, crs="EPSG:4326")
@@ -527,11 +593,11 @@ class GeoSeries(GeoPandasBase, Series):
         from_wkb_or_wkt_function: Callable,
         data,
         index=None,
-        crs: Optional[Any] = None,
+        crs: Any | None = None,
         on_invalid: str = "raise",
         **kwargs,
     ) -> GeoSeries:
-        """Create a GeoSeries from either WKT or WKB values"""
+        """Create a GeoSeries from either WKT or WKB values."""
         if isinstance(data, Series):
             if index is not None:
                 data = data.reindex(index)
@@ -544,8 +610,43 @@ class GeoSeries(GeoPandasBase, Series):
             **kwargs,
         )
 
+    @classmethod
+    def from_arrow(cls, arr, **kwargs) -> GeoSeries:
+        """
+        Construct a GeoSeries from an Arrow array object with a GeoArrow
+        extension type.
+
+        See https://geoarrow.org/ for details on the GeoArrow specification.
+
+        This functions accepts any Arrow array object implementing
+        the `Arrow PyCapsule Protocol`_ (i.e. having an ``__arrow_c_array__``
+        method).
+
+        .. _Arrow PyCapsule Protocol: https://arrow.apache.org/docs/format/CDataInterface/PyCapsuleInterface.html
+
+        .. versionadded:: 1.0
+
+        Parameters
+        ----------
+        arr : pyarrow.Array, Arrow array
+            Any array object implementing the Arrow PyCapsule Protocol
+            (i.e. has an ``__arrow_c_array__`` or ``__arrow_c_stream__``
+            method). The type of the array should be one of the
+            geoarrow geometry types.
+        **kwargs
+            Other parameters passed to the GeoSeries constructor.
+
+        Returns
+        -------
+        GeoSeries
+
+        """
+        from geopandas.io._geoarrow import arrow_to_geometry_array
+
+        return cls(arrow_to_geometry_array(arr), **kwargs)
+
     @property
-    def __geo_interface__(self) -> Dict:
+    def __geo_interface__(self) -> dict:
         """Returns a ``GeoSeries`` as a python feature collection.
 
         Implements the `geo_interface`. The returned python data structure
@@ -555,7 +656,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> from shapely.geometry import Point
         >>> s = geopandas.GeoSeries([Point(1, 1), Point(2, 2), Point(3, 3)])
         >>> s.__geo_interface__
@@ -574,8 +674,8 @@ class GeoSeries(GeoPandasBase, Series):
     def to_file(
         self,
         filename: os.PathLike | typing.IO,
-        driver: Optional[str] = None,
-        index: Optional[bool] = None,
+        driver: str | None = None,
+        index: bool | None = None,
         **kwargs,
     ):
         """Write the ``GeoSeries`` to a file.
@@ -632,7 +732,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> s.to_file('series.shp')  # doctest: +SKIP
 
         >>> s.to_file('series.gpkg', driver='GPKG', layer='name1')  # doctest: +SKIP
@@ -642,7 +741,6 @@ class GeoSeries(GeoPandasBase, Series):
         from geopandas import GeoDataFrame
 
         data = GeoDataFrame({"geometry": self}, index=self.index)
-        data.crs = self.crs
         data.to_file(filename, driver, index=index, **kwargs)
 
     #
@@ -670,9 +768,9 @@ class GeoSeries(GeoPandasBase, Series):
         return _expanddim_logic(df)
 
     def _wrapped_pandas_method(self, mtd, *args, **kwargs):
-        """Wrap a generic pandas method to ensure it returns a GeoSeries"""
+        """Wrap a generic pandas method to ensure it returns a GeoSeries."""
         val = getattr(super(), mtd)(*args, **kwargs)
-        if type(val) == Series:
+        if type(val) is Series:
             val.__class__ = GeoSeries
             val.crs = self.crs
         return val
@@ -689,18 +787,9 @@ class GeoSeries(GeoPandasBase, Series):
         return self._wrapped_pandas_method("take", *args, **kwargs)
 
     @doc(pd.Series)
-    def select(self, *args, **kwargs):
-        return self._wrapped_pandas_method("select", *args, **kwargs)
-
-    @doc(pd.Series)
-    def apply(self, func, convert_dtype: Optional[bool] = None, args=(), **kwargs):
+    def apply(self, func, convert_dtype: bool | None = None, args=(), **kwargs):
         if convert_dtype is not None:
             kwargs["convert_dtype"] = convert_dtype
-        else:
-            # if compat.PANDAS_GE_21 don't pass through, use pandas default
-            # of true to avoid internally triggering the pandas warning
-            if not compat.PANDAS_GE_21:
-                kwargs["convert_dtype"] = True
 
         # to avoid warning
         result = super().apply(func, args=args, **kwargs)
@@ -726,7 +815,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> from shapely.geometry import Polygon
         >>> s = geopandas.GeoSeries(
         ...     [Polygon([(0, 0), (1, 1), (0, 1)]), None, Polygon([])]
@@ -771,7 +859,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Examples
         --------
-
         >>> from shapely.geometry import Polygon
         >>> s = geopandas.GeoSeries(
         ...     [Polygon([(0, 0), (1, 1), (0, 1)]), None, Polygon([])]
@@ -891,7 +978,7 @@ class GeoSeries(GeoPandasBase, Series):
         return super().fillna(value=value, limit=limit, inplace=inplace, **kwargs)
 
     def __contains__(self, other) -> bool:
-        """Allow tests of the form "geom in s"
+        """Allow tests of the form "geom in s".
 
         Tests whether a GeoSeries contains a geometry.
 
@@ -908,7 +995,7 @@ class GeoSeries(GeoPandasBase, Series):
 
     @doc(_explore_geoseries)
     def explore(self, *args, **kwargs):
-        """Interactive map based on folium/leaflet.js"""
+        """Explore with an interactive map based on folium/leaflet.js."""
         return _explore_geoseries(self, *args, **kwargs)
 
     def explode(self, ignore_index=False, index_parts=False) -> GeoSeries:
@@ -955,7 +1042,7 @@ class GeoSeries(GeoPandasBase, Series):
            2    POINT (4 4)
         dtype: geometry
 
-        See also
+        See Also
         --------
         GeoDataFrame.explode
 
@@ -979,20 +1066,24 @@ class GeoSeries(GeoPandasBase, Series):
     @compat.requires_pyproj
     def set_crs(
         self,
-        crs: Optional[Any] = None,
-        epsg: Optional[int] = None,
+        crs: Any | None = None,
+        epsg: int | None = None,
         inplace: bool = False,
         allow_override: bool = False,
     ):
         """
         Set the Coordinate Reference System (CRS) of a ``GeoSeries``.
 
-        NOTE: The underlying geometries are not transformed to this CRS. To
+        Pass ``None`` to remove CRS from the ``GeoSeries``.
+
+        Notes
+        -----
+        The underlying geometries are not transformed to this CRS. To
         transform the geometries to a new CRS, use the ``to_crs`` method.
 
         Parameters
         ----------
-        crs : pyproj.CRS, optional if `epsg` is specified
+        crs : pyproj.CRS | None, optional
             The value can be anything accepted
             by :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
             such as an authority string (eg "EPSG:4326") or a WKT string.
@@ -1060,8 +1151,6 @@ class GeoSeries(GeoPandasBase, Series):
             crs = CRS.from_user_input(crs)
         elif epsg is not None:
             crs = CRS.from_epsg(epsg)
-        else:
-            raise ValueError("Must pass either crs or epsg.")
 
         if not allow_override and self.crs is not None and not self.crs == crs:
             raise ValueError(
@@ -1074,13 +1163,11 @@ class GeoSeries(GeoPandasBase, Series):
             result = self.copy()
         else:
             result = self
-        result.crs = crs
+        result.array.crs = crs
         return result
 
-    def to_crs(
-        self, crs: Optional[Any] = None, epsg: Optional[int] = None
-    ) -> GeoSeries:
-        """Returns a ``GeoSeries`` with all geometries transformed to a new
+    def to_crs(self, crs: Any | None = None, epsg: int | None = None) -> GeoSeries:
+        """Return a ``GeoSeries`` with all geometries transformed to a new
         coordinate reference system.
 
         Transform all geometries in a GeoSeries to a different coordinate
@@ -1160,7 +1247,7 @@ class GeoSeries(GeoPandasBase, Series):
         )
 
     def estimate_utm_crs(self, datum_name: str = "WGS 84"):
-        """Returns the estimated UTM CRS based on the bounds of the dataset.
+        """Return the estimated UTM CRS based on the bounds of the dataset.
 
         .. versionadded:: 0.9
 
@@ -1204,8 +1291,7 @@ class GeoSeries(GeoPandasBase, Series):
         to_wgs84: bool = False,
         **kwargs,
     ) -> str:
-        """
-        Returns a GeoJSON string representation of the GeoSeries.
+        """Return a GeoJSON string representation of the GeoSeries.
 
         Parameters
         ----------
@@ -1255,8 +1341,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         )
 
     def to_wkb(self, hex: bool = False, **kwargs) -> Series:
-        """
-        Convert GeoSeries geometries to WKB
+        r"""Convert GeoSeries geometries to WKB.
 
         Parameters
         ----------
@@ -1272,15 +1357,40 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         Series
             WKB representations of the geometries
 
-        See also
+        See Also
         --------
         GeoSeries.to_wkt
+
+        Examples
+        --------
+        >>> from shapely.geometry import Point, Polygon
+        >>> s = geopandas.GeoSeries(
+        ...     [
+        ...         Point(0, 0),
+        ...         Polygon(),
+        ...         Polygon([(0, 0), (1, 1), (1, 0)]),
+        ...         None,
+        ...     ]
+        ... )
+
+        >>> s.to_wkb()
+        0    b'\x01\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00...
+        1              b'\x01\x03\x00\x00\x00\x00\x00\x00\x00'
+        2    b'\x01\x03\x00\x00\x00\x01\x00\x00\x00\x04\x00...
+        3                                                 None
+        dtype: object
+
+        >>> s.to_wkb(hex=True)
+        0           010100000000000000000000000000000000000000
+        1                                   010300000000000000
+        2    0103000000010000000400000000000000000000000000...
+        3                                                 None
+        dtype: object
         """
         return Series(to_wkb(self.array, hex=hex, **kwargs), index=self.index)
 
     def to_wkt(self, **kwargs) -> Series:
-        """
-        Convert GeoSeries geometries to WKT
+        """Convert GeoSeries geometries to WKT.
 
         Parameters
         ----------
@@ -1308,7 +1418,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         2    POINT (3 3)
         dtype: object
 
-        See also
+        See Also
         --------
         GeoSeries.to_wkb
         """
@@ -1339,7 +1449,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
             struct type.
         include_z : bool, default None
             Only relevant for 'geoarrow' encoding (for WKB, the dimensionality
-            of the individial geometries is preserved).
+            of the individual geometries is preserved).
             If False, return 2D geometries. If True, include the third dimension
             in the output (if a geometry has no third dimension, the z-coordinates
             will be NaN). By default, will infer the dimensionality from the
@@ -1363,7 +1473,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
 
         >>> arrow_array = gser.to_arrow()
         >>> arrow_array
-        <geopandas.io.geoarrow.GeoArrowArray object at ...>
+        <geopandas.io._geoarrow.GeoArrowArray object at ...>
 
         The returned array object needs to be consumed by a library implementing
         the Arrow PyCapsule Protocol. For example, wrapping the data as a
@@ -1379,9 +1489,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         ]
 
         """
-        import pyarrow as pa
-
-        from geopandas.io.geoarrow import (
+        from geopandas.io._geoarrow import (
             GeoArrowArray,
             construct_geometry_array,
             construct_wkb_array,
@@ -1390,9 +1498,6 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
         field_name = self.name if self.name is not None else ""
 
         if geometry_encoding.lower() == "geoarrow":
-            if Version(pa.__version__) < Version("10.0.0"):
-                raise ValueError("Converting to 'geoarrow' requires pyarrow >= 10.0.")
-
             field, geom_arr = construct_geometry_array(
                 np.array(self.array),
                 include_z=include_z,
@@ -1446,7 +1551,7 @@ e": "Feature", "properties": {}, "geometry": {"type": "Point", "coordinates": [3
             Vector data (points, lines, polygons) from `gdf` clipped to
             polygon boundary from mask.
 
-        See also
+        See Also
         --------
         clip : top-level function for clip
 
