@@ -1,5 +1,6 @@
 import warnings
 from contextlib import contextmanager
+from functools import lru_cache
 
 import pandas as pd
 
@@ -22,6 +23,7 @@ def _get_conn(conn_or_engine):
     ----------
     conn_or_engine : Connection or Engine
         A sqlalchemy Connection or Engine instance
+
     Returns
     -------
     Connection
@@ -41,11 +43,12 @@ def _get_conn(conn_or_engine):
         raise ValueError(f"Unknown Connectable: {conn_or_engine}")
 
 
-def _df_to_geodf(df, geom_col="geom", additional_geom_cols=None, crs=None):
-    """
-    Transforms a pandas DataFrame into a GeoDataFrame.
+def _df_to_geodf(df, geom_col="geom", additional_geom_cols=None, crs=None, con=None):
+    """Transform a pandas DataFrame into a GeoDataFrame.
+
     The column 'geom_col' must be a geometry column in WKB representation.
     To be used to convert df based on pd.read_sql to gdf.
+
     Parameters
     ----------
     df : DataFrame
@@ -62,13 +65,15 @@ def _df_to_geodf(df, geom_col="geom", additional_geom_cols=None, crs=None):
         such as an authority string (eg "EPSG:4326") or a WKT string.
         If not set, tries to determine CRS from the SRID associated with the
         first geometry in the database, and assigns that to all geometries.
+    con : sqlalchemy.engine.Connection or sqlalchemy.engine.Engine
+        Active connection to the database to query.
+
     Returns
     -------
     GeoDataFrame
     """
-
     if geom_col not in df:
-        raise ValueError("Query missing geometry column '{}'".format(geom_col))
+        raise ValueError(f"Query missing geometry column '{geom_col}'")
 
     if df.columns.to_list().count(geom_col) > 1:
         raise ValueError(
@@ -100,7 +105,28 @@ def _df_to_geodf(df, geom_col="geom", additional_geom_cols=None, crs=None):
             srid = shapely.get_srid(active_geometry.iat[0])
             # if no defined SRID in geodatabase, returns SRID of 0
             if srid != 0:
-                crs = "epsg:{}".format(srid)
+                try:
+                    spatial_ref_sys_df = _get_spatial_ref_sys_df(con, srid)
+                except pd.errors.DatabaseError:
+                    warning_msg = (
+                        f"Could not find the spatial reference system table "
+                        f"(spatial_ref_sys) in PostGIS."
+                        f"Trying epsg:{srid} as a fallback."
+                    )
+                    warnings.warn(warning_msg, UserWarning, stacklevel=3)
+                    crs = f"epsg:{srid}"
+                else:
+                    if not spatial_ref_sys_df.empty:
+                        auth_name = spatial_ref_sys_df["auth_name"].item()
+                        crs = f"{auth_name}:{srid}"
+                    else:
+                        warning_msg = (
+                            f"Could not find srid {srid} in the "
+                            f"spatial_ref_sys table. "
+                            f"Trying epsg:{srid} as a fallback."
+                        )
+                        warnings.warn(warning_msg, UserWarning, stacklevel=3)
+                        crs = f"epsg:{srid}"
 
     return GeoDataFrame(df, crs=crs, geometry=geom_col)
 
@@ -144,8 +170,7 @@ def _read_postgis(
     params=None,
     chunksize=None,
 ):
-    """
-    Returns a GeoDataFrame corresponding to the result of the query
+    """Return a GeoDataFrame corresponding to the result of the query
     string, which must contain a geometry column in WKB representation.
 
     It is also possible to use :meth:`~GeoDataFrame.read_file` to read from a database.
@@ -195,7 +220,6 @@ def _read_postgis(
     >>> sql = "SELECT ST_AsBinary(geom) AS geom, highway FROM roads"
     >>> df = geopandas.read_postgis(sql, con)  # doctest: +SKIP
     """
-
     if chunksize is None:
         # read all in one chunk and return a single GeoDataFrame
         df = pd.read_sql(
@@ -208,7 +232,7 @@ def _read_postgis(
             chunksize=chunksize,
         )
         return _df_to_geodf(
-            df, geom_col=geom_col, additional_geom_cols=additional_geom_cols, crs=crs
+            df, geom_col=geom_col, additional_geom_cols=additional_geom_cols, crs=crs, con=con
         )
 
     else:
@@ -228,14 +252,16 @@ def _read_postgis(
                 geom_col=geom_col,
                 additional_geom_cols=additional_geom_cols,
                 crs=crs,
+                con=con
             )
             for df in df_generator
         )
 
 
 def _get_geometry_type(gdf):
-    """
-    Get basic geometry type of a GeoDataFrame. See more info from:
+    """Get basic geometry type of a GeoDataFrame.
+
+    See more info from:
     https://geoalchemy-2.readthedocs.io/en/latest/types.html#geoalchemy2.types._GISType
 
     Following rules apply:
@@ -280,16 +306,12 @@ def _get_geometry_type(gdf):
 
 
 def _get_srid_from_crs(gdf):
-    """
-    Get EPSG code from CRS if available. If not, return -1.
-    """
-
+    """Get EPSG code from CRS if available. If not, return 0."""
     # Use geoalchemy2 default for srid
     # Note: undefined srid in PostGIS is 0
     srid = None
     warning_msg = (
-        "Could not parse CRS from the GeoDataFrame. "
-        "Inserting data without defined CRS."
+        "Could not parse CRS from the GeoDataFrame. Inserting data without defined CRS."
     )
     if gdf.crs is not None:
         try:
@@ -307,7 +329,7 @@ def _get_srid_from_crs(gdf):
             warnings.warn(warning_msg, UserWarning, stacklevel=2)
 
     if srid is None:
-        srid = -1
+        srid = 0
         warnings.warn(warning_msg, UserWarning, stacklevel=2)
 
     return srid
@@ -351,11 +373,11 @@ def _psql_insert_copy(tbl, conn, keys, data_iter):
     writer.writerows(data_iter)
     s_buf.seek(0)
 
-    columns = ", ".join('"{}"'.format(k) for k in keys)
+    columns = ", ".join(f'"{k}"' for k in keys)
 
     dbapi_conn = conn.connection
-    sql = 'COPY "{}"."{}" ({}) FROM STDIN WITH CSV'.format(
-        tbl.table.schema, tbl.table.name, columns
+    sql = (
+        f'COPY "{tbl.table.schema}"."{tbl.table.name}" ({columns}) FROM STDIN WITH CSV'
     )
     with dbapi_conn.cursor() as cur:
         # Use psycopg method if it's available
@@ -414,7 +436,6 @@ def _write_postgis(
 
     Examples
     --------
-
     >>> from sqlalchemy import create_engine  # doctest: +SKIP
     >>> engine = create_engine("postgresql://myusername:mypassword@myhost:5432\
 /mydatabase";)  # doctest: +SKIP
@@ -460,18 +481,16 @@ def _write_postgis(
             if connection.dialect.has_table(connection, name, schema):
                 target_srid = connection.execute(
                     text(
-                        "SELECT Find_SRID('{schema}', '{table}', '{geom_col}');".format(
-                            schema=schema_name, table=name, geom_col=geom_name
-                        )
+                        "SELECT Find_SRID(:schema_name, :name, :geom_name);"
+                    ).bindparams(
+                        schema_name=schema_name, name=name, geom_name=geom_name
                     )
                 ).fetchone()[0]
 
                 if target_srid != srid:
                     msg = (
-                        "The CRS of the target table (EPSG:{epsg_t}) differs from the "
-                        "CRS of current GeoDataFrame (EPSG:{epsg_src}).".format(
-                            epsg_t=target_srid, epsg_src=srid
-                        )
+                        f"The CRS of the target table (EPSG:{target_srid}) differs "
+                        f"from the CRS of current GeoDataFrame (EPSG:{srid})."
                     )
                     raise ValueError(msg)
 
@@ -487,3 +506,11 @@ def _write_postgis(
             dtype=dtype,
             method=_psql_insert_copy,
         )
+
+
+@lru_cache
+def _get_spatial_ref_sys_df(con, srid):
+    spatial_ref_sys_sql = (
+        f"SELECT srid, auth_name FROM spatial_ref_sys WHERE srid = {srid}"
+    )
+    return pd.read_sql(spatial_ref_sys_sql, con)
