@@ -2,9 +2,9 @@ import contextlib
 import json
 import os
 import pathlib
-from packaging.version import Version
 
 import numpy as np
+from pandas import ArrowDtype
 
 import shapely
 from shapely import MultiPoint, Point, box
@@ -12,6 +12,7 @@ from shapely import MultiPoint, Point, box
 from geopandas import GeoDataFrame, GeoSeries
 
 import pytest
+from geopandas.testing import assert_geodataframe_equal, assert_geoseries_equal
 
 pytest.importorskip("pyarrow")
 import pyarrow as pa
@@ -22,21 +23,14 @@ DATA_PATH = pathlib.Path(os.path.dirname(__file__)) / "data"
 
 
 def pa_table(table):
-    if Version(pa.__version__) < Version("14.0.0"):
-        return table._pa_table
-    else:
-        return pa.table(table)
+    return pa.table(table)
 
 
 def pa_array(array):
-    if Version(pa.__version__) < Version("14.0.0"):
-        return array._pa_array
-    else:
-        return pa.array(array)
+    return pa.array(array)
 
 
 def assert_table_equal(left, right, check_metadata=True):
-
     geom_type = left["geometry"].type
     # in case of Points (directly the inner fixed_size_list or struct type)
     # -> there are NaNs for empties -> we need to compare them separately
@@ -90,18 +84,15 @@ def assert_table_equal(left, right, check_metadata=True):
 
     if not left.schema.equals(right.schema):
         raise AssertionError(
-            "Schema not equal\nLeft:\n{0}\nRight:\n{1}".format(
-                left.schema, right.schema
-            )
+            f"Schema not equal\nLeft:\n{left.schema}\nRight:\n{right.schema}"
         )
 
     if check_metadata:
         if not left.schema.equals(right.schema, check_metadata=True):
             if not left.schema.metadata == right.schema.metadata:
                 raise AssertionError(
-                    "Metadata not equal\nLeft:\n{0}\nRight:\n{1}".format(
-                        left.schema.metadata, right.schema.metadata
-                    )
+                    f"Metadata not equal\nLeft:\n{left.schema.metadata}\n"
+                    f"Right:\n{right.schema.metadata}"
                 )
         for col in left.schema.names:
             assert left.schema.field(col).equals(
@@ -112,9 +103,7 @@ def assert_table_equal(left, right, check_metadata=True):
         a_left = pa.concat_arrays(left.column(col).chunks)
         a_right = pa.concat_arrays(right.column(col).chunks)
         if not a_left.equals(a_right):
-            raise AssertionError(
-                "Column '{0}' not equal:\n{1}".format(col, a_left.diff(a_right))
-            )
+            raise AssertionError(f"Column '{col}' not equal:\n{a_left.diff(a_right)}")
 
     raise AssertionError("Tables not equal for unknown reason")
 
@@ -154,7 +143,7 @@ def test_geoarrow_export(geometry_type, dim, geometry_encoding, interleaved):
     df["geometry"] = GeoSeries.from_wkb(df["geometry"])
     df["row_number"] = df["row_number"].astype("int32")
     df = GeoDataFrame(df)
-    df.geometry.crs = None
+    df.geometry.array.crs = None
 
     # Read the expected data
     if geometry_encoding == "WKB":
@@ -201,6 +190,20 @@ def test_geoarrow_export(geometry_type, dim, geometry_encoding, interleaved):
 
 
 @pytest.mark.parametrize("encoding", ["WKB", "geoarrow"])
+def test_geoarrow_to_pandas_kwargs(encoding):
+    g = box(0, 0, 10, 10)
+    gdf = GeoDataFrame({"geometry": [g], "i": [1], "s": ["a"]})
+    table = pa_table(gdf.to_arrow(geometry_encoding=encoding))
+    # simulate the `dtype_backend="pyarrow"` option in `pandas.read_parquet`
+    gdf_roundtrip = GeoDataFrame.from_arrow(
+        table, to_pandas_kwargs={"types_mapper": ArrowDtype}
+    )
+    assert isinstance(gdf_roundtrip, GeoDataFrame)
+    assert isinstance(gdf_roundtrip.dtypes["i"], ArrowDtype)
+    assert isinstance(gdf_roundtrip.dtypes["s"], ArrowDtype)
+
+
+@pytest.mark.parametrize("encoding", ["WKB", "geoarrow"])
 def test_geoarrow_multiple_geometry_crs(encoding):
     pytest.importorskip("pyproj")
     # ensure each geometry column has its own crs
@@ -211,11 +214,16 @@ def test_geoarrow_multiple_geometry_crs(encoding):
     meta1 = json.loads(
         result.schema.field("geometry").metadata[b"ARROW:extension:metadata"]
     )
-    assert json.loads(meta1["crs"])["id"]["code"] == 4326
+    assert meta1["crs"]["id"]["code"] == 4326
     meta2 = json.loads(
         result.schema.field("geom2").metadata[b"ARROW:extension:metadata"]
     )
-    assert json.loads(meta2["crs"])["id"]["code"] == 3857
+    assert meta2["crs"]["id"]["code"] == 3857
+
+    roundtripped = GeoDataFrame.from_arrow(result)
+    assert_geodataframe_equal(gdf, roundtripped)
+    assert gdf.geometry.crs == "epsg:4326"
+    assert gdf.geom2.crs == "epsg:3857"
 
 
 @pytest.mark.parametrize("encoding", ["WKB", "geoarrow"])
@@ -233,7 +241,7 @@ def test_geoarrow_series_name_crs(encoding):
         else b"geoarrow.polygon"
     )
     meta = json.loads(field.metadata[b"ARROW:extension:metadata"])
-    assert json.loads(meta["crs"])["id"]["code"] == 4326
+    assert meta["crs"]["id"]["code"] == 4326
 
     # ensure it also works without a name
     gser = GeoSeries([box(0, 0, 10, 10)])
@@ -283,18 +291,6 @@ def test_geoarrow_missing(encoding, interleaved, geom_type):
         geometry=[Point(0, 0) if geom_type == "point" else box(0, 0, 10, 10), None],
         crs="epsg:4326",
     )
-    if (
-        encoding == "geoarrow"
-        and geom_type == "point"
-        and interleaved
-        and Version(pa.__version__) < Version("15.0.0")
-    ):
-        with pytest.raises(
-            ValueError,
-            match="Converting point geometries with missing values is not supported",
-        ):
-            gdf.to_arrow(geometry_encoding=encoding, interleaved=interleaved)
-        return
     result = pa_table(gdf.to_arrow(geometry_encoding=encoding, interleaved=interleaved))
     assert result["geometry"].null_count == 1
     assert result["geometry"].is_null().to_pylist() == [False, True]
@@ -349,7 +345,7 @@ def test_geoarrow_export_with_extension_types(geometry_type, dim):
     df["geometry"] = GeoSeries.from_wkb(df["geometry"])
     df["row_number"] = df["row_number"].astype("int32")
     df = GeoDataFrame(df)
-    df.geometry.crs = None
+    df.geometry.array.crs = None
 
     pytest.importorskip("geoarrow.pyarrow")
 
@@ -362,3 +358,182 @@ def test_geoarrow_export_with_extension_types(geometry_type, dim):
 
         result3 = pa_table(df.to_arrow(geometry_encoding="geoarrow", interleaved=False))
         assert isinstance(result3["geometry"].type, pa.ExtensionType)
+
+
+def test_geoarrow_export_empty():
+    gdf_empty = GeoDataFrame(columns=["col", "geometry"], geometry="geometry")
+    gdf_all_missing = GeoDataFrame(
+        {"col": [1], "geometry": [None]}, geometry="geometry"
+    )
+
+    # no geometries to infer the geometry type -> raise error for now
+    with pytest.raises(NotImplementedError):
+        gdf_empty.to_arrow(geometry_encoding="geoarrow")
+
+    with pytest.raises(NotImplementedError):
+        gdf_all_missing.to_arrow(geometry_encoding="geoarrow")
+
+    # with WKB encoding it roundtrips fine
+    result = pa_table(gdf_empty.to_arrow(geometry_encoding="WKB"))
+    roundtripped = GeoDataFrame.from_arrow(result)
+    assert_geodataframe_equal(gdf_empty, roundtripped)
+
+    result = pa_table(gdf_all_missing.to_arrow(geometry_encoding="WKB"))
+    roundtripped = GeoDataFrame.from_arrow(result)
+    assert_geodataframe_equal(gdf_all_missing, roundtripped)
+
+
+@pytest.mark.parametrize("dim", ["xy", "xyz"])
+@pytest.mark.parametrize(
+    "geometry_type",
+    [
+        "point",
+        "linestring",
+        "polygon",
+        "multipoint",
+        "multilinestring",
+        "multipolygon",
+    ],
+)
+def test_geoarrow_import(geometry_type, dim):
+    base_path = DATA_PATH / "geoarrow"
+    suffix = geometry_type + ("_z" if dim == "xyz" else "")
+
+    # Read the example data
+    df = feather.read_feather(base_path / f"example-{suffix}-wkb.arrow")
+    df["geometry"] = GeoSeries.from_wkb(df["geometry"])
+    df = GeoDataFrame(df)
+    df.geometry.crs = None
+
+    table1 = feather.read_table(base_path / f"example-{suffix}-wkb.arrow")
+    result1 = GeoDataFrame.from_arrow(table1)
+    assert_geodataframe_equal(result1, df)
+
+    table2 = feather.read_table(base_path / f"example-{suffix}-interleaved.arrow")
+    result2 = GeoDataFrame.from_arrow(table2)
+    assert_geodataframe_equal(result2, df)
+
+    table3 = feather.read_table(base_path / f"example-{suffix}.arrow")
+    result3 = GeoDataFrame.from_arrow(table3)
+    assert_geodataframe_equal(result3, df)
+
+
+@pytest.mark.parametrize("encoding", ["WKB", "geoarrow"])
+def test_geoarrow_import_geometry_column(encoding):
+    pytest.importorskip("pyproj")
+    # ensure each geometry column has its own crs
+    gdf = GeoDataFrame(geometry=[box(0, 0, 10, 10)])
+    gdf["centroid"] = gdf.geometry.centroid
+
+    result = GeoDataFrame.from_arrow(pa_table(gdf.to_arrow(geometry_encoding=encoding)))
+    assert_geodataframe_equal(result, gdf)
+    assert result.active_geometry_name == "geometry"
+
+    result = GeoDataFrame.from_arrow(
+        pa_table(gdf[["centroid"]].to_arrow(geometry_encoding=encoding))
+    )
+    assert result.active_geometry_name == "centroid"
+
+    result = GeoDataFrame.from_arrow(
+        pa_table(gdf.to_arrow(geometry_encoding=encoding)), geometry="centroid"
+    )
+    assert result.active_geometry_name == "centroid"
+    assert_geodataframe_equal(result, gdf.set_geometry("centroid"))
+
+
+def test_geoarrow_import_missing_geometry():
+    pytest.importorskip("pyarrow", minversion="14.0.0")
+
+    table = pa.table({"a": [0, 1, 2], "b": [0.1, 0.2, 0.3]})
+    with pytest.raises(ValueError, match="No geometry column found"):
+        GeoDataFrame.from_arrow(table)
+
+    with pytest.raises(ValueError, match="No GeoArrow geometry field found"):
+        GeoSeries.from_arrow(table["a"].chunk(0))
+
+
+def test_geoarrow_import_capsule_interface():
+    # ensure we can import non-pyarrow object
+    pytest.importorskip("pyarrow", minversion="14.0.0")
+    gdf = GeoDataFrame({"col": [1]}, geometry=[box(0, 0, 10, 10)])
+
+    result = GeoDataFrame.from_arrow(gdf.to_arrow())
+    assert_geodataframe_equal(result, gdf)
+
+
+@pytest.mark.parametrize("dim", ["xy", "xyz"])
+@pytest.mark.parametrize(
+    "geometry_type",
+    ["point", "linestring", "polygon", "multipoint", "multilinestring", "multipolygon"],
+)
+def test_geoarrow_import_from_extension_types(geometry_type, dim):
+    # ensure the exported data can be imported by geoarrow-pyarrow and are
+    # recognized as extension types
+    pytest.importorskip("pyproj")
+    base_path = DATA_PATH / "geoarrow"
+    suffix = geometry_type + ("_z" if dim == "xyz" else "")
+
+    # Read the example data
+    df = feather.read_feather(base_path / f"example-{suffix}-wkb.arrow")
+    df["geometry"] = GeoSeries.from_wkb(df["geometry"])
+    df = GeoDataFrame(df, crs="EPSG:3857")
+
+    pytest.importorskip("geoarrow.pyarrow")
+
+    with with_geoarrow_extension_types():
+        result1 = GeoDataFrame.from_arrow(
+            pa_table(df.to_arrow(geometry_encoding="WKB"))
+        )
+        assert_geodataframe_equal(result1, df)
+
+        result2 = GeoDataFrame.from_arrow(
+            pa_table(df.to_arrow(geometry_encoding="geoarrow"))
+        )
+        assert_geodataframe_equal(result2, df)
+
+        result3 = GeoDataFrame.from_arrow(
+            pa_table(df.to_arrow(geometry_encoding="geoarrow", interleaved=False))
+        )
+        assert_geodataframe_equal(result3, df)
+
+
+def test_geoarrow_import_geoseries():
+    pytest.importorskip("pyproj")
+    gp = pytest.importorskip("geoarrow.pyarrow")
+    ser = GeoSeries.from_wkt(["POINT (1 1)", "POINT (2 2)"], crs="EPSG:3857")
+
+    with with_geoarrow_extension_types():
+        arr = gp.array(ser.to_arrow(geometry_encoding="WKB"))
+        result = GeoSeries.from_arrow(arr)
+        assert_geoseries_equal(result, ser)
+
+        arr = gp.array(ser.to_arrow(geometry_encoding="geoarrow"))
+        result = GeoSeries.from_arrow(arr)
+        assert_geoseries_equal(result, ser)
+
+        # the name is lost when going through a pyarrow.Array
+        ser.name = "name"
+        arr = gp.array(ser.to_arrow())
+        result = GeoSeries.from_arrow(arr)
+        assert result.name is None
+        # we can specify the name as one of the kwargs
+        result = GeoSeries.from_arrow(arr, name="test")
+        assert_geoseries_equal(result, ser)
+
+
+def test_geoarrow_import_unknown_geoarrow_type():
+    gdf = GeoDataFrame({"col": [1]}, geometry=[box(0, 0, 10, 10)])
+    table = pa_table(gdf.to_arrow())
+    schema = table.schema
+    new_field = schema.field("geometry").with_metadata(
+        {
+            b"ARROW:extension:name": b"geoarrow.unknown",
+            b"ARROW:extension:metadata": b"{}",
+        }
+    )
+
+    new_schema = pa.schema([schema.field(0), new_field])
+    new_table = table.cast(new_schema)
+
+    with pytest.raises(TypeError, match="Unknown GeoArrow extension type"):
+        GeoDataFrame.from_arrow(new_table)
