@@ -6,19 +6,25 @@ see geopandas.tests.util for more information.
 """
 
 import os
+import uuid
 import warnings
 from importlib.util import find_spec
 
 import pandas as pd
 
 import geopandas
-import geopandas._compat as compat
 from geopandas import GeoDataFrame, read_file, read_postgis
+from geopandas._compat import HAS_PYPROJ
 from geopandas.io.sql import _get_conn as get_conn
 from geopandas.io.sql import _write_postgis as write_postgis
 
 import pytest
-from geopandas.tests.util import create_postgis, create_spatialite, validate_boro_df
+from geopandas.tests.util import (
+    create_postgis,
+    create_spatialite,
+    mock,
+    validate_boro_df,
+)
 
 try:
     from sqlalchemy import text
@@ -267,10 +273,8 @@ class TestIO:
         out_geom = "the_geom"
         create_postgis(con, df_nybb, geom_col=orig_geom)
 
-        sql = """SELECT borocode, boroname, shape_leng, shape_area,
-                    {} as {} FROM nybb;""".format(
-            orig_geom, out_geom
-        )
+        sql = f"""SELECT borocode, boroname, shape_leng, shape_area,
+                    {orig_geom} as {out_geom} FROM nybb;"""
         df = read_postgis(sql, con, geom_col=out_geom)
 
         validate_boro_df(df)
@@ -327,11 +331,11 @@ class TestIO:
         """Tests that geometry with NULL is accepted."""
         con = connection_spatialite
         geom_col = df_nybb.geometry.name
-        df_nybb.geometry.iat[0] = None
+        df_nybb.loc[0, "geometry"] = None
         create_spatialite(con, df_nybb)
         sql = (
             "SELECT ogc_fid, borocode, boroname, shape_leng, shape_area, "
-            'AsEWKB("{0}") AS "{0}" FROM nybb'.format(geom_col)
+            f'AsEWKB("{geom_col}") AS "{geom_col}" FROM nybb'
         )
         df = read_postgis(sql, con, geom_col=geom_col)
         validate_boro_df(df)
@@ -343,7 +347,7 @@ class TestIO:
         create_spatialite(con, df_nybb)
         sql = (
             "SELECT ogc_fid, borocode, boroname, shape_leng, shape_area, "
-            'ST_AsBinary("{0}") AS "{0}" FROM nybb'.format(geom_col)
+            f'ST_AsBinary("{geom_col}") AS "{geom_col}" FROM nybb'
         )
         df = read_postgis(sql, con, geom_col=geom_col)
         validate_boro_df(df)
@@ -497,7 +501,7 @@ class TestIO:
         # Write to db
         write_postgis(df_nybb, con=engine, name=table, if_exists="fail")
         # Validate
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -513,7 +517,7 @@ class TestIO:
         # Write to db
         write_postgis(df_nybb, con=engine, name=table, if_exists="fail")
         # Validate
-        sql = text('SELECT * FROM "{table}";'.format(table=table))
+        sql = text(f'SELECT * FROM "{table}";')
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -529,7 +533,7 @@ class TestIO:
             # Write to db
             write_postgis(df_nybb, con=con, name=table, if_exists="fail")
             # Validate
-            sql = text("SELECT * FROM {table};".format(table=table))
+            sql = text(f"SELECT * FROM {table};")
             df = read_postgis(sql, con, geom_col="geometry")
             validate_boro_df(df)
 
@@ -567,7 +571,7 @@ class TestIO:
         # Overwrite
         write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
         # Validate
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -585,21 +589,17 @@ class TestIO:
         write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
         write_postgis(df_nybb, con=engine, name=table, if_exists="append")
         # Validate
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         new_rows, new_cols = df.shape
 
         # There should be twice as many rows in the new table
         assert new_rows == orig_rows * 2, (
-            "There should be {target} rows,found: {current}".format(
-                target=orig_rows * 2, current=new_rows
-            ),
+            f"There should be {orig_rows * 2} rows,found: {new_rows}",
         )
         # Number of columns should stay the same
         assert new_cols == orig_cols, (
-            "There should be {target} columns,found: {current}".format(
-                target=orig_cols, current=new_cols
-            ),
+            f"There should be {orig_cols} columns,found: {new_cols}",
         )
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
@@ -612,7 +612,7 @@ class TestIO:
         table = "nybb"
 
         # Write to db
-        df_nybb.crs = None
+        df_nybb.geometry.array.crs = None
         with pytest.warns(UserWarning, match="Could not parse CRS from the GeoDataF"):
             write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
         # Validate that srid is -1
@@ -623,7 +623,7 @@ class TestIO:
         )
         with engine.connect() as conn:
             target_srid = conn.execute(sql).fetchone()[0]
-        assert target_srid == 0, "SRID should be 0, found %s" % target_srid
+        assert target_srid == 0, f"SRID should be 0, found {target_srid}"
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
     def test_write_postgis_with_esri_authority(self, engine_postgis, df_nybb):
@@ -646,7 +646,7 @@ class TestIO:
         )
         with engine.connect() as conn:
             target_srid = conn.execute(sql).fetchone()[0]
-        assert target_srid == 102003, "SRID should be 102003, found %s" % target_srid
+        assert target_srid == 102003, f"SRID should be 102003, found {target_srid}"
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
     def test_write_postgis_geometry_collection(
@@ -662,14 +662,10 @@ class TestIO:
         write_postgis(df_geom_collection, con=engine, name=table, if_exists="replace")
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             geom_type = conn.execute(sql).fetchone()[0]
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
 
         assert geom_type.upper() == "GEOMETRYCOLLECTION"
@@ -691,11 +687,7 @@ class TestIO:
         )
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             res = conn.execute(sql).fetchall()
         assert res[0][0].upper() == "LINESTRING"
@@ -714,11 +706,7 @@ class TestIO:
         write_postgis(df_linear_ring, con=engine, name=table, if_exists="replace")
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT(GeometryType(geometry)) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             geom_type = conn.execute(sql).fetchone()[0]
 
@@ -741,17 +729,13 @@ class TestIO:
             chunksize=1,
         )
         # Validate row count
-        sql = text("SELECT COUNT(geometry) FROM {table};".format(table=table))
+        sql = text(f"SELECT COUNT(geometry) FROM {table};")
         with engine.connect() as conn:
             row_cnt = conn.execute(sql).fetchone()[0]
         assert row_cnt == 3
 
         # Validate geometry type
-        sql = text(
-            "SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;".format(
-                table=table
-            )
-        )
+        sql = text(f"SELECT DISTINCT GeometryType(geometry) FROM {table} ORDER BY 1;")
         with engine.connect() as conn:
             res = conn.execute(sql).fetchall()
         assert res[0][0].upper() == "LINESTRING"
@@ -767,7 +751,7 @@ class TestIO:
 
         table = "nybb"
         schema_to_use = "test"
-        sql = text("CREATE SCHEMA IF NOT EXISTS {schema};".format(schema=schema_to_use))
+        sql = text(f"CREATE SCHEMA IF NOT EXISTS {schema_to_use};")
         with engine.begin() as conn:
             conn.execute(sql)
 
@@ -775,9 +759,7 @@ class TestIO:
             df_nybb, con=engine, name=table, if_exists="replace", schema=schema_to_use
         )
         # Validate
-        sql = text(
-            "SELECT * FROM {schema}.{table};".format(schema=schema_to_use, table=table)
-        )
+        sql = text(f"SELECT * FROM {schema_to_use}.{table};")
 
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
@@ -793,7 +775,7 @@ class TestIO:
 
         table = "nybb"
         schema_to_use = "test"
-        sql = text("CREATE SCHEMA IF NOT EXISTS {schema};".format(schema=schema_to_use))
+        sql = text(f"CREATE SCHEMA IF NOT EXISTS {schema_to_use};")
         with engine.begin() as conn:
             conn.execute(sql)
 
@@ -802,11 +784,7 @@ class TestIO:
                 df_nybb, con=engine, name=table, if_exists="fail", schema=schema_to_use
             )
             # Validate
-            sql = text(
-                "SELECT * FROM {schema}.{table};".format(
-                    schema=schema_to_use, table=table
-                )
-            )
+            sql = text(f"SELECT * FROM {schema_to_use}.{table};")
 
             df = read_postgis(sql, engine, geom_col="geometry")
             validate_boro_df(df)
@@ -820,9 +798,7 @@ class TestIO:
             df_nybb, con=engine, name=table, if_exists="replace", schema=schema_to_use
         )
         # Validate
-        sql = text(
-            "SELECT * FROM {schema}.{table};".format(schema=schema_to_use, table=table)
-        )
+        sql = text(f"SELECT * FROM {schema_to_use}.{table};")
 
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
@@ -839,7 +815,7 @@ class TestIO:
         write_postgis(df_3D_geoms, con=engine, name=table, if_exists="replace")
 
         # Check that all geometries have 3 dimensions
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         assert list(df.geometry.has_z) == [True, True, True]
 
@@ -856,7 +832,7 @@ class TestIO:
         write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
 
         # Check that the row order matches
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         assert df["BoroCode"].tolist() == correct_order
 
@@ -874,7 +850,7 @@ class TestIO:
         write_postgis(df_nybb, con=engine, name=table, if_exists="append")
 
         # Check that the row order matches
-        sql = text("SELECT * FROM {table};".format(table=table))
+        sql = text(f"SELECT * FROM {table};")
         df = read_postgis(sql, engine, geom_col="geometry")
         validate_boro_df(df)
 
@@ -896,10 +872,22 @@ class TestIO:
             write_postgis(df_nybb2, con=engine, name=table, if_exists="append")
 
     @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
-    @pytest.mark.xfail(
-        compat.PANDAS_GE_20 and not compat.PANDAS_GE_202,
-        reason="Duplicate columns are dropped in read_sql with pandas 2.0.0 and 2.0.1",
-    )
+    def test_append_without_crs(self, engine_postgis, df_nybb):
+        # This test was included in #3328 when the default value for no
+        # CRS was changed from an SRID of -1 to 0. This resolves issues
+        # of appending dataframes to postgis that have no CRS as postgis
+        # no CRS value is 0.
+        engine = engine_postgis
+        df_nybb = df_nybb.set_crs(None, allow_override=True)
+        table = "nybb"
+
+        write_postgis(df_nybb, con=engine, name=table, if_exists="replace")
+        # append another dataframe with no crs
+
+        df_nybb2 = df_nybb
+        write_postgis(df_nybb2, con=engine, name=table, if_exists="append")
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
     def test_duplicate_geometry_column_fails(self, engine_postgis):
         """
         Tests that a ValueError is raised if an SQL query returns two geometry columns.
@@ -910,3 +898,324 @@ class TestIO:
 
         with pytest.raises(ValueError):
             read_postgis(sql, engine, geom_col="geom")
+
+    @pytest.mark.parametrize("connection_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_read_non_epsg_crs(self, connection_postgis, df_nybb):
+        con = connection_postgis
+        df_nybb = df_nybb.to_crs(crs="esri:54052")
+        create_postgis(con, df_nybb, srid=54052)
+
+        sql = "SELECT * FROM nybb;"
+        df = read_postgis(sql, con)
+        validate_boro_df(df)
+        assert df.crs == "ESRI:54052"
+
+    @pytest.mark.skipif(not HAS_PYPROJ, reason="pyproj not installed")
+    @mock.patch("shapely.get_srid")
+    @pytest.mark.parametrize("connection_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_read_srid_not_in_table(self, mock_get_srid, connection_postgis, df_nybb):
+        # mock a non-existent srid for edge case if shapely has an srid
+        # not present in postgis table.
+        pyproj = pytest.importorskip("pyproj")
+
+        mock_get_srid.return_value = 99999
+
+        con = connection_postgis
+        df_nybb = df_nybb.to_crs(crs="epsg:4326")
+        create_postgis(con, df_nybb)
+
+        sql = "SELECT * FROM nybb;"
+        with pytest.raises(pyproj.exceptions.CRSError, match="crs not found"):
+            with pytest.warns(UserWarning, match="Could not find srid 99999"):
+                read_postgis(sql, con)
+
+    @mock.patch("geopandas.io.sql._get_spatial_ref_sys_df")
+    @pytest.mark.parametrize("connection_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_read_no_spatial_ref_sys_table_in_postgis(
+        self, mock_get_spatial_ref_sys_df, connection_postgis, df_nybb
+    ):
+        # mock for a non-existent spatial_ref_sys database
+
+        mock_get_spatial_ref_sys_df.side_effect = pd.errors.DatabaseError
+
+        con = connection_postgis
+        df_nybb = df_nybb.to_crs(crs="epsg:4326")
+        create_postgis(con, df_nybb, srid=4326)
+
+        sql = "SELECT * FROM nybb;"
+        with pytest.warns(
+            UserWarning, match="Could not find the spatial reference system table"
+        ):
+            df = read_postgis(sql, con)
+
+        assert df.crs == "EPSG:4326"
+
+    @pytest.mark.parametrize("connection_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_read_non_epsg_crs_chunksize(self, connection_postgis, df_nybb):
+        """Test chunksize argument with non epsg crs"""
+        chunksize = 2
+        con = connection_postgis
+        df_nybb = df_nybb.to_crs(crs="esri:54052")
+
+        create_postgis(con, df_nybb, srid=54052)
+
+        sql = "SELECT * FROM nybb;"
+        df = pd.concat(read_postgis(sql, con, chunksize=chunksize))
+
+        validate_boro_df(df)
+        assert df.crs == "ESRI:54052"
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_to_postgis_append_with_custom_geometry_column_name(
+        self, engine_postgis, df_nybb
+    ):
+        """
+        Tests that appending works correctly with custom geometry column names.
+        This ensures the SQL injection fix doesn't break normal functionality.
+        """
+        engine = engine_postgis
+        table = "test_custom_geom"
+
+        # Create GeoDataFrame with custom geometry column name
+        gdf = df_nybb.copy()
+        gdf = gdf.rename_geometry("custom_geom")
+
+        # Write initial data
+        write_postgis(gdf, con=engine, name=table, if_exists="replace")
+
+        # Append more data - this should work without SQL injection vulnerability
+        write_postgis(gdf, con=engine, name=table, if_exists="append")
+
+        # Verify data was appended correctly
+        sql = text(f'SELECT * FROM "{table}";')
+        df = read_postgis(sql, engine, geom_col="custom_geom")
+        assert len(df) == len(gdf) * 2
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_to_postgis_sql_injection_prevention_geometry_column(
+        self, engine_postgis, df_nybb
+    ):
+        """
+        Tests that SQL injection attempts via geometry column name are prevented.
+        This test verifies that malicious geometry column names are treated as
+        literal strings and not executed as SQL code.
+
+        This test is based on issue #3679 which reported SQL injection vulnerability
+        in the Find_SRID query when using to_postgis with if_exists='append'.
+        """
+        engine = engine_postgis
+        table = "test_sql_injection"
+
+        # Create a test table first to ensure append mode triggers
+        # the vulnerable code path
+        gdf_normal = df_nybb.copy()
+        write_postgis(gdf_normal, con=engine, name=table, if_exists="replace")
+
+        # Create a test table that would be targeted by SQL injection
+        # If SQL injection works, this table would be dropped
+        # Use random name to avoid conflicts with existing tables
+        test_target_table = f"test_target_{uuid.uuid4().hex[:12]}"
+        with engine.begin() as conn:
+            conn.execute(
+                text(f"CREATE TABLE IF NOT EXISTS {test_target_table} (id INTEGER);")
+            )
+            conn.execute(text(f"INSERT INTO {test_target_table} (id) VALUES (1);"))
+
+        # Create GeoDataFrame with malicious geometry column name
+        # This simulates an attacker trying to inject SQL via geometry column name
+        gdf_malicious = df_nybb.copy()
+        malicious_geom_name = f"geom'; DROP TABLE {test_target_table}; --"
+        gdf_malicious = gdf_malicious.rename_geometry(malicious_geom_name)
+
+        # First, verify both tables exist
+        sql_check = text(f"SELECT COUNT(*) FROM {table};")
+        sql_check_target = text(f"SELECT COUNT(*) FROM {test_target_table};")
+        with engine.connect() as conn:
+            initial_count = conn.execute(sql_check).fetchone()[0]
+            target_exists = conn.execute(sql_check_target).fetchone()[0] == 1
+
+        assert target_exists, "Target table should exist before SQL injection attempt"
+
+        # Attempt append - this should fail because the geometry column name
+        # doesn't match, but it should NOT execute the injected SQL
+        try:
+            write_postgis(gdf_malicious, con=engine, name=table, if_exists="append")
+            # If it doesn't raise an error, that's also fine - the important
+            # thing is that SQL injection didn't occur
+        except (ValueError, Exception):
+            # Expected - the geometry column name doesn't exist in the table
+            pass
+
+        # Verify both tables still exist and weren't dropped by SQL injection
+        with engine.connect() as conn:
+            final_count = conn.execute(sql_check).fetchone()[0]
+            target_still_exists = conn.execute(sql_check_target).fetchone()[0] == 1
+
+        assert final_count == initial_count, (
+            "Test table should still exist after SQL injection attempt"
+        )
+        assert target_still_exists, (
+            f"Target table '{test_target_table}' should not be dropped by SQL injection"
+        )
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_to_postgis_sql_injection_prevention_geometry_column_union(
+        self, engine_postgis, df_nybb
+    ):
+        """
+        Tests another SQL injection pattern: UNION SELECT attack
+        via geometry column name.
+        """
+        engine = engine_postgis
+        table = "test_sql_injection_union"
+
+        # Create initial table
+        gdf_normal = df_nybb.copy()
+        write_postgis(gdf_normal, con=engine, name=table, if_exists="replace")
+
+        # Create a test table that would be targeted by UNION SELECT injection
+        # Use random name to avoid conflicts with existing tables
+        test_target_table = f"test_target_{uuid.uuid4().hex[:12]}"
+        with engine.begin() as conn:
+            conn.execute(
+                text(f"CREATE TABLE IF NOT EXISTS {test_target_table} (password TEXT);")
+            )
+            conn.execute(
+                text(f"INSERT INTO {test_target_table} (password) VALUES ('secret');")
+            )
+
+        # Create GeoDataFrame with UNION-based SQL injection attempt
+        gdf_malicious = df_nybb.copy()
+        malicious_geom_name = (
+            f"geom') UNION SELECT password FROM {test_target_table} WHERE '1'='1"
+        )
+        gdf_malicious = gdf_malicious.rename_geometry(malicious_geom_name)
+
+        # Verify tables exist before injection attempt
+        sql_check = text(f"SELECT COUNT(*) FROM {table};")
+        sql_check_target = text(f"SELECT COUNT(*) FROM {test_target_table};")
+        with engine.connect() as conn:
+            initial_count = conn.execute(sql_check).fetchone()[0]
+            target_exists = conn.execute(sql_check_target).fetchone()[0] == 1
+
+        assert target_exists, "Target table should exist before SQL injection attempt"
+
+        # Attempt append - should not execute the UNION SELECT
+        try:
+            write_postgis(gdf_malicious, con=engine, name=table, if_exists="append")
+        except (ValueError, Exception):
+            # Expected - geometry column name mismatch
+            pass
+
+        # Verify table integrity maintained
+        with engine.connect() as conn:
+            final_count = conn.execute(sql_check).fetchone()[0]
+            target_still_exists = conn.execute(sql_check_target).fetchone()[0] == 1
+
+        assert final_count == initial_count, (
+            "Test table should still exist after SQL injection attempt"
+        )
+        assert target_still_exists, (
+            f"Target table '{test_target_table}' should not be "
+            "affected by SQL injection"
+        )
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_to_postgis_sql_injection_prevention_geometry_column_comment(
+        self, engine_postgis, df_nybb
+    ):
+        """
+        Tests SQL injection via comment syntax in geometry column name.
+        """
+        engine = engine_postgis
+        table = "test_sql_injection_comment"
+
+        # Create initial table
+        gdf_normal = df_nybb.copy()
+        write_postgis(gdf_normal, con=engine, name=table, if_exists="replace")
+
+        # Create GeoDataFrame with comment-based SQL injection
+        gdf_malicious = df_nybb.copy()
+        malicious_geom_name = "geom' -- DROP TABLE test_sql_injection_comment;"
+        gdf_malicious = gdf_malicious.rename_geometry(malicious_geom_name)
+
+        # Verify table exists
+        sql_check = text(f"SELECT COUNT(*) FROM {table};")
+        with engine.connect() as conn:
+            initial_count = conn.execute(sql_check).fetchone()[0]
+
+        # Attempt append
+        try:
+            write_postgis(gdf_malicious, con=engine, name=table, if_exists="append")
+        except (ValueError, Exception):
+            pass
+
+        # Verify table still exists
+        with engine.connect() as conn:
+            final_count = conn.execute(sql_check).fetchone()[0]
+            assert final_count == initial_count
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_to_postgis_sql_injection_prevention_geometry_column_semicolon(
+        self, engine_postgis, df_nybb
+    ):
+        """
+        Tests SQL injection via semicolon in geometry column name.
+        """
+        engine = engine_postgis
+        table = "test_sql_injection_semicolon"
+
+        # Create initial table
+        gdf_normal = df_nybb.copy()
+        write_postgis(gdf_normal, con=engine, name=table, if_exists="replace")
+
+        # Create GeoDataFrame with semicolon-based SQL injection
+        gdf_malicious = df_nybb.copy()
+        malicious_geom_name = "geom; DELETE FROM test_sql_injection_semicolon;"
+        gdf_malicious = gdf_malicious.rename_geometry(malicious_geom_name)
+
+        # Verify table exists
+        sql_check = text(f"SELECT COUNT(*) FROM {table};")
+        with engine.connect() as conn:
+            initial_count = conn.execute(sql_check).fetchone()[0]
+
+        # Attempt append
+        try:
+            write_postgis(gdf_malicious, con=engine, name=table, if_exists="append")
+        except (ValueError, Exception):
+            pass
+
+        # Verify table still exists and data is intact
+        with engine.connect() as conn:
+            final_count = conn.execute(sql_check).fetchone()[0]
+            assert final_count == initial_count
+
+    @pytest.mark.parametrize("engine_postgis", POSTGIS_DRIVERS, indirect=True)
+    def test_to_postgis_append_with_special_characters_in_geometry_column(
+        self, engine_postgis, df_nybb
+    ):
+        """
+        Tests that legitimate geometry column names with special characters
+        (but not SQL injection) still work correctly.
+        This ensures the fix doesn't break normal functionality with valid
+        but unusual column names.
+        """
+        engine = engine_postgis
+        table = "test_special_chars"
+
+        # Create GeoDataFrame with geometry column name containing special characters
+        # (but not SQL injection patterns)
+        gdf = df_nybb.copy()
+        # Use a valid but unusual geometry column name
+        gdf = gdf.rename_geometry("geom_with_underscores_123")
+
+        # Write initial data
+        write_postgis(gdf, con=engine, name=table, if_exists="replace")
+
+        # Append more data - should work correctly
+        write_postgis(gdf, con=engine, name=table, if_exists="append")
+
+        # Verify data was appended correctly
+        sql = text(f'SELECT * FROM "{table}";')
+        df = read_postgis(sql, engine, geom_col="geom_with_underscores_123")
+        assert len(df) == len(gdf) * 2
