@@ -1,4 +1,4 @@
-"""Vendored, cut down version of pyogrio/util.py for use with fiona"""
+"""Vendored, cut down version of pyogrio/util.py for use with fiona."""
 
 import re
 import sys
@@ -6,11 +6,7 @@ from urllib.parse import urlparse
 
 
 def vsi_path(path: str) -> str:
-    """
-    Ensure path is a local path or a GDAL-compatible vsi path.
-
-    """
-
+    """Ensure path is a local path or a GDAL-compatible vsi path."""
     # path is already in GDAL format
     if path.startswith("/vsi"):
         return path
@@ -56,8 +52,7 @@ CURLSCHEMES = {k for k, v in SCHEMES.items() if v == "curl"}
 
 
 def _parse_uri(path: str):
-    """
-    Parse a URI
+    """Parse a URI.
 
     Returns a tuples of (path, archive, scheme)
 
@@ -92,8 +87,7 @@ def _parse_uri(path: str):
 
 
 def _construct_vsi_path(path, archive, scheme) -> str:
-    """Convert a parsed path to a GDAL VSI path"""
-
+    """Convert a parsed path to a GDAL VSI path."""
     prefix = ""
     suffix = ""
     schemes = scheme.split("+")
@@ -102,9 +96,7 @@ def _construct_vsi_path(path, archive, scheme) -> str:
         schemes.insert(0, "zip")
 
     if schemes:
-        prefix = "/".join(
-            "vsi{0}".format(SCHEMES[p]) for p in schemes if p and p != "file"
-        )
+        prefix = "/".join(f"vsi{SCHEMES[p]}" for p in schemes if p and p != "file")
 
         if schemes[-1] in CURLSCHEMES:
             suffix = f"{schemes[-1]}://"
@@ -113,6 +105,6 @@ def _construct_vsi_path(path, archive, scheme) -> str:
         if archive:
             return "/{}/{}{}/{}".format(prefix, suffix, archive, path.lstrip("/"))
         else:
-            return "/{}/{}{}".format(prefix, suffix, path)
+            return f"/{prefix}/{suffix}{path}"
 
     return path
