@@ -8,7 +8,7 @@ from shapely.geometry import GeometryCollection, LineString, Point, Polygon, box
 
 import geopandas
 from geopandas import GeoDataFrame, GeoSeries, overlay, read_file
-from geopandas._compat import HAS_PYPROJ, PANDAS_GE_20
+from geopandas._compat import HAS_PYPROJ, PANDAS_INFER_STR
 
 import pytest
 from geopandas.testing import assert_geodataframe_equal, assert_geoseries_equal
@@ -78,10 +78,8 @@ def test_overlay(dfs_index, how):
     # construction of result
 
     def _read(name):
-        expected = read_file(
-            os.path.join(DATA, "polys", "df1_df2-{0}.geojson".format(name))
-        )
-        expected.crs = None
+        expected = read_file(os.path.join(DATA, "polys", f"df1_df2-{name}.geojson"))
+        expected.geometry.array.crs = None
         for col in expected.columns[expected.dtypes == "int32"]:
             expected[col] = expected[col].astype("int64")
         return expected
@@ -92,7 +90,6 @@ def test_overlay(dfs_index, how):
         expected = pd.concat(
             [expected_intersection, expected_difference], ignore_index=True, sort=False
         )
-        expected["col1"] = expected["col1"].astype(float)
     else:
         expected = _read(how)
 
@@ -110,6 +107,22 @@ def test_overlay(dfs_index, how):
         result = result.reset_index(drop=True)
         expected = _read("difference-inverse")
         assert_geodataframe_equal(result, expected, check_column_type=False)
+
+
+@pytest.mark.parametrize("keep_geom_type", [True, False])
+def test_overlay_invalid_input(keep_geom_type):
+    invalid_polygon = Polygon([(0, 0), (1, 1), (1, 2), (1, 1), (0, 0)])
+    square = Polygon(((0, 0), (1, 0), (1, 1), (0, 1), (0, 0)))
+    df1 = geopandas.GeoDataFrame({"A": [1, 2]}, geometry=[invalid_polygon, square])
+    df2 = geopandas.GeoDataFrame({"B": [5, 6]}, geometry=[square, square])
+
+    df1_df2 = df1.overlay(df2, keep_geom_type=keep_geom_type)
+    # As overlay only supports inputs with uniform geometry types, `make_valid` of
+    # `invalid_polygon` should always result in POLYGON EMPTY, so the number of results
+    # will should always be 2.
+    assert len(df1_df2) == 2
+    if keep_geom_type:
+        assert df1_df2.geom_type.isin(["Polygon"]).all()
 
 
 @pytest.mark.filterwarnings("ignore:GeoSeries crs mismatch:UserWarning")
@@ -146,9 +159,7 @@ def test_overlay_nybb(how, nybb_filename):
         # read union one, further down below we take the appropriate subset
         expected = read_file(os.path.join(DATA, "nybb_qgis", "qgis-union.shp"))
     else:
-        expected = read_file(
-            os.path.join(DATA, "nybb_qgis", "qgis-{0}.shp".format(how))
-        )
+        expected = read_file(os.path.join(DATA, "nybb_qgis", f"qgis-{how}.shp"))
 
     # The result of QGIS for 'union' contains incorrect geometries:
     # 24 is a full original circle overlapping with unioned geometries, and
@@ -162,6 +173,10 @@ def test_overlay_nybb(how, nybb_filename):
 
     if how == "identity":
         expected = expected[expected.BoroCode.notnull()].copy()
+        boro_code_dtype = result["BoroCode"].dtype
+        if boro_code_dtype in ("int32", "int64"):
+            # Depending on the pandas version, the dtype might be int32 or int64
+            expected["BoroCode"] = expected["BoroCode"].astype(boro_code_dtype)
 
     # Order GeoDataFrames
     expected = expected.sort_values(cols).reset_index(drop=True)
@@ -260,7 +275,7 @@ def test_overlay_overlap(how):
         raise pytest.skip()
 
     expected = read_file(
-        os.path.join(DATA, "overlap", "df1_df2_overlap-{0}.geojson".format(how))
+        os.path.join(DATA, "overlap", f"df1_df2_overlap-{how}.geojson")
     )
 
     if how == "union":
@@ -371,7 +386,7 @@ def test_crs_mismatch(dfs, how):
 
 
 def test_empty_intersection(dfs):
-    df1, df2 = dfs
+    df1, _df2 = dfs
     polys3 = GeoSeries(
         [
             Polygon([(-1, -1), (-3, -1), (-3, -3), (-1, -3)]),
@@ -386,7 +401,7 @@ def test_empty_intersection(dfs):
 
 def test_correct_index(dfs):
     # GH883 - case where the index was not properly reset
-    df1, df2 = dfs
+    _df1, df2 = dfs
     polys3 = GeoSeries(
         [
             Polygon([(1, 1), (3, 1), (3, 3), (1, 3)]),
@@ -405,7 +420,7 @@ def test_correct_index(dfs):
 
 
 def test_warn_on_keep_geom_type(dfs):
-    df1, df2 = dfs
+    _df1, df2 = dfs
     polys3 = GeoSeries(
         [
             Polygon([(1, 1), (3, 1), (3, 3), (1, 3)]),
@@ -506,7 +521,7 @@ def test_overlay_strict(how, keep_geom_type, geom_types):
             os.path.join(
                 DATA,
                 "strict",
-                "{t}_{h}_{s}.geojson".format(t=geom_types, h=how, s=keep_geom_type),
+                f"{geom_types}_{how}_{keep_geom_type}.geojson",
             )
         )
 
@@ -751,28 +766,26 @@ def test_empty_overlay_return_non_duplicated_columns(nybb_filename):
 def test_non_overlapping(how):
     p1 = Polygon([(0, 0), (2, 0), (2, 2), (0, 2)])
     p2 = Polygon([(3, 3), (5, 3), (5, 5), (3, 5)])
-    df1 = GeoDataFrame({"col1": [1], "geometry": [p1]})
-    df2 = GeoDataFrame({"col2": [2], "geometry": [p2]})
+    df1 = GeoDataFrame({"id": [1], "col1": [1], "geometry": [p1]})
+    df2 = GeoDataFrame({"id": [2], "col2": [2], "geometry": [p2]})
     result = overlay(df1, df2, how=how)
 
     if how == "intersection":
-        if PANDAS_GE_20:
-            index = None
-        else:
-            index = pd.Index([], dtype="object")
-
         expected = GeoDataFrame(
             {
+                "id_1": np.array([], dtype="int64"),
                 "col1": np.array([], dtype="int64"),
+                "id_2": np.array([], dtype="int64"),
                 "col2": np.array([], dtype="int64"),
                 "geometry": [],
-            },
-            index=index,
+            }
         )
     elif how == "union":
         expected = GeoDataFrame(
             {
+                "id_1": [1, np.nan],
                 "col1": [1, np.nan],
+                "id_2": [np.nan, 2],
                 "col2": [np.nan, 2],
                 "geometry": [p1, p2],
             }
@@ -780,7 +793,9 @@ def test_non_overlapping(how):
     elif how == "identity":
         expected = GeoDataFrame(
             {
-                "col1": [1.0],
+                "id_1": [1],
+                "col1": [1],
+                "id_2": [np.nan],
                 "col2": [np.nan],
                 "geometry": [p1],
             }
@@ -788,7 +803,9 @@ def test_non_overlapping(how):
     elif how == "symmetric_difference":
         expected = GeoDataFrame(
             {
+                "id_1": [1, np.nan],
                 "col1": [1, np.nan],
+                "id_2": [np.nan, 2],
                 "col2": [np.nan, 2],
                 "geometry": [p1, p2],
             }
@@ -796,6 +813,7 @@ def test_non_overlapping(how):
     elif how == "difference":
         expected = GeoDataFrame(
             {
+                "id": [1],
                 "col1": [1],
                 "geometry": [p1],
             }
@@ -811,60 +829,92 @@ def test_no_intersection():
     gdf2 = GeoDataFrame({"bar": ["1", "3", "5"]}, geometry=gs.translate(1))
 
     expected = GeoDataFrame(columns=["foo", "bar", "geometry"])
+    if PANDAS_INFER_STR:
+        expected = expected.astype({"foo": "str", "bar": "str"})
     result = overlay(gdf1, gdf2, how="intersection")
     assert_geodataframe_equal(result, expected, check_index_type=False)
 
 
+def test_zero_len():
+    # https://github.com/geopandas/geopandas/issues/3422
+    gdf1 = GeoDataFrame(
+        {
+            "geometry": [
+                Polygon([[0.0, 0.0], [2.0, 0.0], [2.0, 2.0], [0.0, 2.0], [0.0, 0.0]])
+            ]
+        },
+        crs=4326,
+    )
+    # overlay with empty geodataframe shouldn't throw
+    gdf2 = GeoDataFrame({"geometry": []}, crs=4326)
+    res = gdf1.overlay(gdf2, how="union")
+    assert_geodataframe_equal(res, gdf1)
+
+    gdf2 = GeoDataFrame(geometry=[], crs=4326)
+    res = gdf1.overlay(gdf2, how="union")
+    assert_geodataframe_equal(res, gdf1)
+
+
 class TestOverlayWikiExample:
     def setup_method(self):
-        self.layer_a = GeoDataFrame(geometry=[box(0, 2, 6, 6)])
+        self.layer_a = GeoDataFrame(data={"id": ["a"]}, geometry=[box(0, 2, 6, 6)])
 
-        self.layer_b = GeoDataFrame(geometry=[box(4, 0, 10, 4)])
+        self.layer_b = GeoDataFrame(data={"id": ["b"]}, geometry=[box(4, 0, 10, 4)])
 
-        self.intersection = GeoDataFrame(geometry=[box(4, 2, 6, 4)])
+        self.intersection = GeoDataFrame(
+            data={"id_1": ["a"], "id_2": ["b"]}, geometry=[box(4, 2, 6, 4)]
+        )
 
         self.union = GeoDataFrame(
+            data={"id_1": ["a", "a", np.nan], "id_2": ["b", np.nan, "b"]},
             geometry=[
                 box(4, 2, 6, 4),
                 Polygon([(4, 2), (0, 2), (0, 6), (6, 6), (6, 4), (4, 4), (4, 2)]),
                 Polygon([(10, 0), (4, 0), (4, 2), (6, 2), (6, 4), (10, 4), (10, 0)]),
-            ]
+            ],
         )
 
         self.a_difference_b = GeoDataFrame(
-            geometry=[Polygon([(4, 2), (0, 2), (0, 6), (6, 6), (6, 4), (4, 4), (4, 2)])]
+            data={"id": ["a"]},
+            geometry=[
+                Polygon([(4, 2), (0, 2), (0, 6), (6, 6), (6, 4), (4, 4), (4, 2)])
+            ],
         )
 
         self.b_difference_a = GeoDataFrame(
+            data={"id": ["b"]},
             geometry=[
                 Polygon([(10, 0), (4, 0), (4, 2), (6, 2), (6, 4), (10, 4), (10, 0)])
-            ]
+            ],
         )
 
         self.symmetric_difference = GeoDataFrame(
+            data={"id_1": ["a", np.nan], "id_2": [np.nan, "b"]},
             geometry=[
                 Polygon([(4, 2), (0, 2), (0, 6), (6, 6), (6, 4), (4, 4), (4, 2)]),
                 Polygon([(10, 0), (4, 0), (4, 2), (6, 2), (6, 4), (10, 4), (10, 0)]),
-            ]
+            ],
         )
 
         self.a_identity_b = GeoDataFrame(
+            data={"id_1": ["a", "a"], "id_2": ["b", np.nan]},
             geometry=[
                 box(4, 2, 6, 4),
                 Polygon([(4, 2), (0, 2), (0, 6), (6, 6), (6, 4), (4, 4), (4, 2)]),
-            ]
+            ],
         )
 
         self.b_identity_a = GeoDataFrame(
+            data={"id_1": ["b", "b"], "id_2": ["a", np.nan]},
             geometry=[
                 box(4, 2, 6, 4),
                 Polygon([(10, 0), (4, 0), (4, 2), (6, 2), (6, 4), (10, 4), (10, 0)]),
-            ]
+            ],
         )
 
     def test_intersection(self):
         df_result = overlay(self.layer_a, self.layer_b, how="intersection")
-        assert df_result.geom_equals(self.intersection).all()
+        assert_geodataframe_equal(df_result, self.intersection)
 
     def test_union(self):
         df_result = overlay(self.layer_a, self.layer_b, how="union")
