@@ -1,6 +1,5 @@
 import itertools
 import warnings
-from packaging.version import Version
 
 import numpy as np
 import pandas as pd
@@ -29,7 +28,7 @@ matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 
 try:  # skipif and importorskip do not work for decorators
-    from matplotlib.testing.decorators import check_figures_equal
+    from matplotlib.testing.decorators import check_figures_equal, image_comparison
 
     MPL_DECORATORS = True
 except ImportError:
@@ -221,13 +220,6 @@ class TestPointPlotting:
                 np.linspace(0, 0.0, 1.0, self.N), ax.collections[0].get_alpha()
             )
 
-    # TODO temporary skip for mpl 3.8.0.dev
-    # (https://github.com/matplotlib/matplotlib/issues/25162)
-    @pytest.mark.skipif(
-        Version(matplotlib.__version__) >= Version("3.8.0.dev")
-        and Version(matplotlib.__version__) < Version("3.8.0"),
-        reason="failing with matplotlib dev",
-    )
     def test_legend(self):
         with warnings.catch_warnings(record=True) as _:  # don't print warning
             # legend ignored if color is given.
@@ -281,7 +273,7 @@ class TestPointPlotting:
         actual_colors_orig = ax.collections[0].get_facecolors()
         exp_colors = cmap(np.arange(10) / (20))
         np.testing.assert_array_equal(exp_colors, actual_colors_orig)
-        fig, ax = plt.subplots()
+        _fig, ax = plt.subplots()
         self.df[1:].plot(column="values", ax=ax, norm=norm, cmap=cmap)
         actual_colors_sub = ax.collections[0].get_facecolors()
         np.testing.assert_array_equal(actual_colors_orig[1], actual_colors_sub[0])
@@ -420,7 +412,7 @@ class TestPointPlotting:
             self.df.plot(column="cats_object", categories="non_list")
 
         with pytest.raises(
-            ValueError, match="Column contains values not listed in categories."
+            ValueError, match="Column contains values not listed in categories"
         ):
             self.df.plot(column="cats_object", categories=["cat1"])
 
@@ -459,6 +451,17 @@ class TestPointPlotting:
         df["category"] = df["values"].astype("str")
         df.plot("category", missing_kwds={"facecolor": "none"}, legend=True)
 
+    def test_missing_aspect(self):
+        self.df.loc[0, "values"] = np.nan
+        ax = self.df.plot(
+            "values",
+            missing_kwds={"color": "r"},
+            categorical=True,
+            legend=True,
+            aspect=2,
+        )
+        assert ax.get_aspect() == 2
+
 
 class TestPointZPlotting:
     def setup_method(self):
@@ -493,6 +496,28 @@ class TestLineStringPlotting:
             index=list("ABCDEFGHIJ"),
         )
         self.df3 = GeoDataFrame({"geometry": self.linearrings, "values": values})
+
+    def test_autolim_false(self):
+        """Test linestring plot preserving axes limits."""
+        ax = self.lines[: self.N // 2].plot()
+        ylim = ax.get_ylim()
+        self.lines.plot(ax=ax, autolim=False)
+        assert ax.get_ylim() == ylim
+        ax = self.df[: self.N // 2].plot()
+        ylim = ax.get_ylim()
+        self.df.plot(ax=ax, autolim=False)
+        assert ax.get_ylim() == ylim
+
+    def test_autolim_true(self):
+        """Test linestring plot autoscaling axes limits."""
+        ax = self.lines[: self.N // 2].plot()
+        ylim = ax.get_ylim()
+        self.lines.plot(ax=ax, autolim=True)
+        assert ax.get_ylim() != ylim
+        ax = self.df[: self.N // 2].plot()
+        ylim = ax.get_ylim()
+        self.df.plot(ax=ax, autolim=True)
+        assert ax.get_ylim() != ylim
 
     def test_single_color(self):
         ax = self.lines.plot(color="green")
@@ -598,7 +623,7 @@ class TestLineStringPlotting:
         actual_colors_orig = ax.collections[0].get_edgecolors()
         exp_colors = cmap(np.arange(10) / (20))
         np.testing.assert_array_equal(exp_colors, actual_colors_orig)
-        fig, ax = plt.subplots()
+        _fig, ax = plt.subplots()
         self.df[1:].plot(column="values", ax=ax, norm=norm, cmap=cmap)
         actual_colors_sub = ax.collections[0].get_edgecolors()
         np.testing.assert_array_equal(actual_colors_orig[1], actual_colors_sub[0])
@@ -636,6 +661,28 @@ class TestPolygonPlotting:
         t3 = Polygon([(2, 0), (3, 0), (3, 1)])
         df_nan = GeoDataFrame({"geometry": t3, "values": [np.nan]})
         self.df3 = pd.concat([self.df, df_nan])
+
+    def test_autolim_false(self):
+        """Test polygon plot preserving axes limits."""
+        ax = self.polys[:1].plot()
+        xlim = ax.get_xlim()
+        self.polys.plot(ax=ax, autolim=False)
+        assert ax.get_xlim() == xlim
+        ax = self.df[:1].plot()
+        xlim = ax.get_xlim()
+        self.df.plot(ax=ax, autolim=False)
+        assert ax.get_xlim() == xlim
+
+    def test_autolim_true(self):
+        """Test polygon plot autoscaling axes limits."""
+        ax = self.polys[:1].plot()
+        xlim = ax.get_xlim()
+        self.polys.plot(ax=ax, autolim=True)
+        assert ax.get_xlim() != xlim
+        ax = self.df[:1].plot()
+        xlim = ax.get_xlim()
+        self.df.plot(ax=ax, autolim=True)
+        assert ax.get_xlim() != xlim
 
     def test_single_color(self):
         ax = self.polys.plot(color="green")
@@ -793,6 +840,47 @@ class TestPolygonPlotting:
         cax = _get_colorbar_ax(ax.get_figure())
         assert cax.get_xlabel() == label_txt
 
+    if MPL_DECORATORS:
+        """Test that geometries are properly normalized so holes appear."""
+
+        @image_comparison(
+            ["polygon_with_holes"],
+            extensions=["png", "pdf"],
+            remove_text=True,
+            savefig_kwarg={"dpi": 300, "bbox_inches": "tight"},
+        )
+        def test_plot_polygon_with_holes(self):
+            geoms = [
+                Polygon(
+                    [(0, 0), (0, 5), (5, 5), (5, 0)],
+                    [
+                        [(1, 1), (1, 2), (2, 2), (2, 1)],
+                        [(3, 2), (3, 3), (4, 3), (4, 2)],
+                    ],
+                )
+            ]
+
+            _df = GeoDataFrame(geometry=geoms)
+            _df.plot()
+    else:
+
+        def test_plot_polygon_with_holes(self):
+            geoms = [
+                Polygon(
+                    [(0, 0), (0, 5), (5, 5), (5, 0)],
+                    [
+                        [(1, 1), (1, 2), (2, 2), (2, 1)],
+                        [(3, 2), (3, 3), (4, 3), (4, 2)],
+                    ],
+                )
+            ]
+
+            _df = GeoDataFrame(geometry=geoms)
+            ax = _df.plot()
+            plotted_vertices = ax.collections[0].get_paths()[0].vertices
+            expected_vertices = _df.normalize().get_coordinates().to_numpy()
+            np.testing.assert_array_equal(plotted_vertices, expected_vertices)
+
     def test_fmt_ignore(self):
         # test if fmt is removed if scheme is not passed (it would raise Error)
         # GH #1253
@@ -864,6 +952,38 @@ class TestPolygonPlotting:
                 [0.7, 0.7, 0.2, 0.2], ax.collections[0].get_alpha()
             )
 
+    if MPL_DECORATORS:
+        """Test multipolygons properly orient so holes will appear.
+        Test partly derived from shapely PR #1933."""
+
+        @image_comparison(
+            ["multipolygon_with_holes"],
+            extensions=["png", "pdf"],
+            remove_text=True,
+            savefig_kwarg={"dpi": 300, "bbox_inches": "tight"},
+        )
+        def test_multipolygons_with_interior(self):
+            poly1 = box(0, 0, 1, 1).difference(box(0.2, 0.2, 0.5, 0.5))
+            poly2 = box(3, 3, 6, 6).difference(box(4, 4, 5, 5))
+            multipoly = MultiPolygon([poly1, poly2])
+            _df = GeoDataFrame(geometry=[multipoly])
+            _df.plot()
+    else:
+
+        def test_multipolygons_with_interior(self):
+            poly1 = box(0, 0, 1, 1).difference(box(0.2, 0.2, 0.5, 0.5))
+            poly2 = box(3, 3, 6, 6).difference(box(4, 4, 5, 5))
+            multipoly = MultiPolygon([poly1, poly2])
+            _df = GeoDataFrame(geometry=[multipoly])
+            ax = _df.plot()
+            plotted_vertices = np.append(
+                ax.collections[0].get_paths()[0].vertices,
+                ax.collections[0].get_paths()[1].vertices,
+                axis=0,
+            )
+            expected_vertices = _df.normalize().get_coordinates().to_numpy()
+            np.testing.assert_array_equal(plotted_vertices, expected_vertices)
+
     def test_subplots_norm(self):
         # colors of subplots are the same as for plot (norm is applied)
         cmap = matplotlib.cm.viridis_r
@@ -872,7 +992,7 @@ class TestPolygonPlotting:
         actual_colors_orig = ax.collections[0].get_facecolors()
         exp_colors = cmap(np.arange(2) / (10))
         np.testing.assert_array_equal(exp_colors, actual_colors_orig)
-        fig, ax = plt.subplots()
+        _fig, ax = plt.subplots()
         self.df[1:].plot(column="values", ax=ax, norm=norm, cmap=cmap)
         actual_colors_sub = ax.collections[0].get_facecolors()
         np.testing.assert_array_equal(actual_colors_orig[1], actual_colors_sub[0])
@@ -1466,7 +1586,7 @@ class TestPlotCollections:
 
         from geopandas.plotting import _plot_point_collection
 
-        fig, ax = plt.subplots()
+        _fig, ax = plt.subplots()
         coll = _plot_point_collection(ax, self.points)
         assert isinstance(coll, PathCollection)
         ax.cla()
@@ -1535,7 +1655,7 @@ class TestPlotCollections:
 
         from geopandas.plotting import _plot_linestring_collection
 
-        fig, ax = plt.subplots()
+        _fig, ax = plt.subplots()
         coll = _plot_linestring_collection(ax, self.lines)
         assert isinstance(coll, LineCollection)
         ax.cla()
@@ -1619,7 +1739,7 @@ class TestPlotCollections:
 
         from geopandas.plotting import _plot_polygon_collection
 
-        fig, ax = plt.subplots()
+        _fig, ax = plt.subplots()
         coll = _plot_polygon_collection(ax, self.polygons)
         assert isinstance(coll, PatchCollection)
         ax.cla()
@@ -1799,6 +1919,8 @@ def test_column_values():
     t2 = Polygon([(1, 0), (2, 0), (2, 1)])
     polys = GeoSeries([t1, t2], index=list("AB"))
     df = GeoDataFrame({"geometry": polys, "values": [0, 1]})
+    numeric_index_polys = GeoSeries([t1, t2], index=[0, 1])
+    numeric_index_df = GeoDataFrame({"geometry": numeric_index_polys, "values": [0, 1]})
 
     # Test with continuous values
     ax = df.plot(column="values")
@@ -1817,6 +1939,11 @@ def test_column_values():
     colors_series = ax.collections[0].get_facecolors()
     np.testing.assert_array_equal(colors, colors_series)
     ax = df.plot(column=df["values"].values, categorical=True)
+    colors_array = ax.collections[0].get_facecolors()
+    np.testing.assert_array_equal(colors, colors_array)
+
+    # Test with pd.Index
+    ax = numeric_index_df.plot(column=numeric_index_df.index, categorical=True)
     colors_array = ax.collections[0].get_facecolors()
     np.testing.assert_array_equal(colors, colors_array)
 
@@ -1875,13 +2002,13 @@ def _check_colors(N, actual_colors, expected_colors, alpha=None):
     actual_colors = map(tuple, actual_colors)
     all_actual_colors = list(itertools.islice(itertools.cycle(actual_colors), N))
 
-    assert len(all_actual_colors) == len(
-        expected_colors
-    ), "Different lengths of actual and expected colors!"
+    assert len(all_actual_colors) == len(expected_colors), (
+        "Different lengths of actual and expected colors!"
+    )
 
     for actual, expected in zip(all_actual_colors, expected_colors):
-        assert actual == conv.to_rgba(expected, alpha=alpha), "{} != {}".format(
-            actual, conv.to_rgba(expected, alpha=alpha)
+        assert actual == conv.to_rgba(expected, alpha=alpha), (
+            f"{actual} != {conv.to_rgba(expected, alpha=alpha)}"
         )
 
 
@@ -1911,7 +2038,7 @@ def _get_ax(fig, label):
     for ax in fig.axes:
         if ax.get_label() == label:
             return ax
-    raise ValueError("no ax found with label {0}".format(label))
+    raise ValueError(f"no ax found with label {label}")
 
 
 def _get_colorbar_ax(fig):
