@@ -1,4 +1,4 @@
-from packaging.version import Version
+import warnings
 from statistics import mean
 
 import numpy as np
@@ -64,7 +64,7 @@ def _explore(
     map_kwds={},
     **kwargs,
 ):
-    """Interactive map based on GeoPandas and folium/leaflet.js
+    """Explore data in interactive map based on GeoPandas and folium/leaflet.js.
 
     Generate an interactive leaflet map based on :class:`~geopandas.GeoDataFrame`
 
@@ -203,6 +203,13 @@ def _explore(
     highlight_kwds : dict (default {})
         Style to be passed to folium highlight_function. Uses the same keywords
         as ``style_kwds``. When empty, defaults to ``{"fillOpacity": 0.75}``.
+    missing_kwds : dict (default {})
+        Additional style for missing values:
+
+        color : str
+            Color of missing values. Defaults to ``None``, which uses Folium's default.
+        label : str (default "NaN")
+            Legend entry for missing values.
     tooltip_kwds : dict (default {})
         Additional keywords to be passed to :class:`folium.features.GeoJsonTooltip`,
         e.g. ``aliases``, ``labels``, or ``sticky``.
@@ -269,38 +276,31 @@ def _explore(
     """
 
     def _colormap_helper(_cmap, n_resample=None, idx=None):
-        """Helper for MPL deprecation - GH#2596"""
+        """Return the color map specified.
+
+        Helper function for MPL deprecation - GH#2596.
+        """
         if not n_resample:
             return cm.get_cmap(_cmap)
         else:
-            if MPL_361:
-                return cm.get_cmap(_cmap).resampled(n_resample)(idx)
-            else:
-                return cm.get_cmap(_cmap, n_resample)(idx)
+            return cm.get_cmap(_cmap).resampled(n_resample)(idx)
 
     try:
         import re
 
         import branca as bc
         import folium
-        import matplotlib
         import matplotlib.pyplot as plt
         from mapclassify import classify
+        from matplotlib import colormaps as cm
         from matplotlib import colors
-
-        # isolate MPL version - GH#2596
-        MPL_361 = Version(matplotlib.__version__) >= Version("3.6.1")
-        if MPL_361:
-            from matplotlib import colormaps as cm
-        else:
-            from matplotlib import cm
 
     except (ImportError, ModuleNotFoundError):
         raise ImportError(
-            "The 'folium', 'matplotlib' and 'mapclassify' packages are required for "
-            "'explore()'. You can install them using "
-            "'conda install -c conda-forge folium matplotlib mapclassify' "
-            "or 'pip install folium matplotlib mapclassify'."
+            "The 'folium>=0.12', 'matplotlib' and 'mapclassify' packages "
+            "are required for 'explore()'. You can install them using "
+            "'conda install -c conda-forge \"folium>=0.12\" matplotlib mapclassify' "
+            "or 'pip install \"folium>=0.12\" matplotlib mapclassify'."
         )
 
     # xyservices is an optional dependency
@@ -336,12 +336,17 @@ def _explore(
     if len(json_not_supported_cols) > 0:
         gdf = gdf.astype({c: "string" for c in json_not_supported_cols})
 
+    if not isinstance(gdf.index, pd.MultiIndex) and (
+        is_datetime64_any_dtype(gdf.index) or (gdf.index.dtype == "object")
+    ):
+        gdf.index = gdf.index.astype("string")
+
     # create folium.Map object
     if m is None:
         # Get bounds to specify location and map extent
         bounds = gdf.total_bounds
         location = kwargs.pop("location", None)
-        if location is None:
+        if location is None and not np.isnan(bounds).all():
             x = mean([bounds[0], bounds[2]])
             y = mean([bounds[1], bounds[3]])
             location = (y, x)
@@ -393,6 +398,15 @@ def _explore(
         # fit bounds to get a proper zoom level
         if fit:
             m.fit_bounds([[bounds[1], bounds[0]], [bounds[3], bounds[2]]])
+
+    if gdf.is_empty.all():
+        warnings.warn(
+            "The GeoSeries you are attempting to plot is "
+            "composed of empty geometries. Nothing has been displayed.",
+            UserWarning,
+            stacklevel=3,
+        )
+        return m
 
     for map_kwd in _MAP_KWARGS:
         kwargs.pop(map_kwd, None)
@@ -635,13 +649,13 @@ def _explore(
         ~(gdf.geometry.isna() | gdf.geometry.is_empty)  # drop missing or empty geoms
     ].__geo_interface__
     for feature in feature_collection["features"]:
-        for k in feature["properties"]:
+        for prop in feature["properties"]:
             # escape the curly braces in values
-            if isinstance(feature["properties"][k], str):
-                feature["properties"][k] = re.sub(
+            if isinstance(feature["properties"][prop], str):
+                feature["properties"][prop] = re.sub(
                     r"\{{2,}",
                     lambda x: "{% raw %}" + x.group(0) + "{% endraw %}",
-                    feature["properties"][k],
+                    feature["properties"][prop],
                 )
 
     # add dataframe to map
@@ -747,7 +761,7 @@ def _explore(
 
 
 def _tooltip_popup(type, fields, gdf, **kwds):
-    """get tooltip or popup"""
+    """Get tooltip or popup."""
     import folium
 
     # specify fields to show in the tooltip
@@ -774,8 +788,7 @@ def _tooltip_popup(type, fields, gdf, **kwds):
 
 
 def _categorical_legend(m, title, categories, colors):
-    """
-    Add categorical legend to a map
+    """Add categorical legend to a map.
 
     The implementation is using the code originally written by Michel Metran
     (@michelmetran) and released on GitHub
@@ -794,7 +807,6 @@ def _categorical_legend(m, title, categories, colors):
     colors : list-like
         list of colors (in the same order as categories)
     """
-
     # Header to Add
     head = """
     {% macro header(this, kwargs) %}
@@ -906,7 +918,7 @@ def _explore_geoseries(
     map_kwds={},
     **kwargs,
 ):
-    """Interactive map based on GeoPandas and folium/leaflet.js
+    """Interactive map based on GeoPandas and folium/leaflet.js.
 
     Generate an interactive leaflet map based on :class:`~geopandas.GeoSeries`
 
